requirements.txt
streamlit==1.37.1
langchain==0.2.12
langchain-community==0.2.11
langchain-openai==0.1.22
chromadb==0.5.3
psycopg2-binary==2.9.9
tiktoken==0.7.0